face swap/requirements.txt
opencv-python
dlib
numpy
pillow
streamlit

